Initial commit!
Basic functionality for scanning for cookies, pixel images, HTML
validity, and local/session storage. Also includes a module that
incrementally adds new URLs to scan, the "crawl" module - allows one to
scan whole sites. The origin for the crawled URLs must match the same
origin as the page the links are found on.

The cookie module needs an input function to accept JS cookies - it is
expected to be located at `input/cookies.ts` and it must have a default
export, a function that programmatically accepts cookies. The other
modules do not have input.